+++ src/pytest_recording/exceptions.py
@@ -1,0 +1,4 @@
+class UsageError(Exception):
+    """Error in plugin usage."""
+
+    __module__ = "builtins"